str2udouble: max as double

+++ src/tool_getparam.c
@@ -808,7 +808,7 @@ ParameterError getparameter(const char *flag, /* f or -long-flag */
         break;
       case 'c': /* connect-timeout */
         err = str2udouble(&config->connecttimeout, nextarg,
-                          LONG_MAX/1000);
+                          (double)LONG_MAX/1000);
         if(err)
           return err;
         break;
@@ -1374,7 +1374,8 @@ ParameterError getparameter(const char *flag, /* f or -long-flag */
           return err;
         break;
       case 'R': /* --expect100-timeout */
-        err = str2udouble(&config->expect100timeout, nextarg, LONG_MAX/1000);
+        err = str2udouble(&config->expect100timeout, nextarg,
+                          (double)LONG_MAX/1000);
         if(err)
           return err;
         break;
@@ -2067,7 +2068,7 @@ ParameterError getparameter(const char *flag, /* f or -long-flag */
       break;
     case 'm':
       /* specified max time */
-      err = str2udouble(&config->timeout, nextarg, LONG_MAX/1000);
+      err = str2udouble(&config->timeout, nextarg, (double)LONG_MAX/1000);
       if(err)
         return err;
       break;
